backend/requirements.txt
fastapi
uvicorn 
pydantic 
scikit-learn 
pandas 
joblib 
sqlalchemy 
python-dotenv
tqdm
sqlalchemy
pyodbc
python-multipart


